requirements.txt
python-can
pytest
matplotlib
dash
pyyaml
